Changed datatype of sideLength from int to double
Keeping in line with trying to return a double value for area() and perimeter()

+++ Lab1/src/Quadrilateral.java
@@ -1,4 +1,4 @@
 public abstract class Quadrilateral implements Polygon {
     protected int sides;
-    public int[] sideLength = new int[sides];
+    public double[] sideLength = new double[sides];
 }
